src/main/java/yaho/service/ProductService.java
package yaho.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import yaho.domain.Product;
import yaho.repository.ProductRepository;

import java.util.List;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class ProductService {
    private final ProductRepository productRepository;

    @Transactional
    public void add(Product product) {
        productRepository.save(product);
    }

    public List<Product>list(){
        return productRepository.findAll();
    }

    public Product findByName(String productName) {
        return productRepository.findByName(productName);
    }
}
